libmain: add a compile option to avoid saving wifi params to flash
Add source code for sdk_wifi_param_save_protect() and a compile time,
WIFI_PARAM_SAVE, option to skip writing the wifi state to flash. This
avoids wear on the flash and does not appear to be necessary when the
app initializes the state anyway.

Define WIFI_PARAM_SAVE to 0 for the wificfg example.